Moved snapshot tests to their own file, and added template testing
